requirements.txt
# Usahawan Muda - Backend Dependencies
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
psycopg2-binary>=2.9,<3.0
Pillow>=10.0,<11.0
django-filter>=23.5,<24.0

# Production
gunicorn>=21.2,<22.0
whitenoise>=6.6,<7.0
dj-database-url>=2.1,<3.0
cloudinary>=1.36,<2.0
django-cloudinary-storage>=0.3,<1.0